pyproject.toml
[project]
name = "demo-app"
version = "1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
